moves util methods to single file
