+++ src/desktop/apps/articles/routes.ts
@@ -1,5 +1,5 @@
 import { stitch as _stitch } from "@artsy/stitch"
-import { areThirdPartyAdsEnabled } from "desktop/apps/article/third_party_ads_enabled"
+import { areThirdPartyAdsEnabled } from "desktop/apps/article/helpers"
 import { getCurrentUnixTimestamp } from "reaction/Components/Publishing/Constants"
 import App from "desktop/apps/articles/components/App"
 import {
